requirements.txt
pyTelegramBotAPI
pytest
jproperties
flake8
matplotlib
coverage
pytest-mock
python-telegram-bot-calendar